fixed reforging
